auth fixed
