+++ src/app/auth/layout.tsx
@@ -1,0 +1,69 @@
+'use client'
+
+import { SITE_NAME } from '@/constants/seo.constants'
+import { Grid, CssBaseline, Paper, Link, Typography } from '@mui/material'
+
+function Copyright(props: any) {
+	return (
+		<Typography
+			variant='body2'
+			color='text.secondary'
+			align='center'
+			{...props}
+		>
+			{'Copyright © '}
+			<Link
+				color='inherit'
+				href='#'
+			>
+				{SITE_NAME}
+			</Link>{' '}
+			{new Date().getFullYear()}
+			{'.'}
+		</Typography>
+	)
+}
+
+export default function AuthLayout({
+	children
+}: Readonly<{
+	children: React.ReactNode
+}>) {
+	return (
+		<Grid
+			container
+			component='main'
+			sx={{ height: '100vh' }}
+		>
+			<CssBaseline />
+			<Grid
+				item
+				xs={false}
+				sm={4}
+				md={7}
+				sx={{
+					backgroundImage: 'url(/AuthFon.jpeg)',
+					backgroundRepeat: 'no-repeat',
+					backgroundColor: t =>
+						t.palette.mode === 'light'
+							? t.palette.grey[50]
+							: t.palette.grey[900],
+					backgroundSize: 'cover',
+					backgroundPosition: 'center'
+				}}
+			/>
+			<Grid
+				item
+				xs={12}
+				sm={8}
+				md={5}
+				component={Paper}
+				elevation={6}
+				square
+			>
+				{children}
+				<Copyright sx={{ mt: 3 }} />
+			</Grid>
+		</Grid>
+	)
+}
