+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next'
-import { Noto_Sans } from 'next/font/google'
+// import { Noto_Sans } from 'next/font/google'
 import './globals.css'
 import { SITE_NAME } from '@/constants/seo.constants'
 import { Providers } from './providers'
